Waveform backward compatibility (#22)
This PR adds NRAR-style backward compatibility to the waveform
objects, and to the object returned when requesting multiple
extrapolations.  Closes #20.

It also solves a problem with downloading the catalog.  We now
explicitly choose temp files inside the cache directory to use
during download and zipping of the catalog.  This ensures that the
temporary files are on the same disk as the final result, which
avoids copying and bugs related to moving across filesystems.
Closes #16.