Update for new `spherical` syntax (#40)
This is a #major release.  Although this package's user interface is essentially unchanged, the `spherical` package has broken backwards compatibility, which means that some closely associated tasks will look very different.  The main change is that `spherical` can now accurately (and quickly) compute Wigner D matrices and spin-weighted spherical harmonics to very large ell values.  Specifically, ell~30 (which is already higher than evaluation of the usual expressions) was the previous limit before overflow would start to appear for certain arguments.  The new expressions in `spherical` should be valid for ell in the thousands, which can be important for calculations involving large displacements.